requirements.txt
annotated-types==0.7.0
black==25.11.0
blinker==1.9.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
decorator==4.4.2
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.2.4
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
instagramy==4.5
instagrapi==1.17.0
isort==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
moviepy==1.0.3
mypy_extensions==1.1.0
numpy==2.3.5
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
proglog==0.1.12
pycryptodomex==3.17
pydantic==1.10.5
pydantic_core==2.33.2
PySocks==1.7.1
python-dotenv==1.2.1
pytokens==0.3.0
requests==2.32.4
SQLAlchemy==2.0.44
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
